@@ -20,16 +20,16 @@
     }
   },
   "require": {
-    "php": "^7.1|^8",
+    "php": "^8.1",
     "ext-gmp": "*",
     "ext-json": "*",
     "ext-openssl": "*",
     "paragonie/easy-ecc": "^1",
     "paragonie/hidden-string": "^1|^2",
-    "paragonie/paseto": ">=2.2"
+    "paragonie/paseto": "^3"
   },
   "require-dev": {
-    "phpunit/phpunit": "^7|^8|^9",
+    "phpunit/phpunit": "^9",
     "vimeo/psalm": "^4"
   },
   "scripts": {
