@@ -3,8 +3,6 @@ declare(strict_types=1);
 namespace ParagonIE\Paserk;
 
 use ParagonIE\Paseto\Protocol\{
-    Version1,
-    Version2,
     Version3,
     Version4
 };
@@ -28,18 +26,11 @@ class Util
      */
     public static function getPasetoVersion(string $version): ProtocolInterface
     {
-        switch ($version) {
-            case 'k1':
-                return new Version1();
-            case 'k2':
-                return new Version2();
-            case 'k3':
-                return new Version3();
-            case 'k4':
-                return new Version4();
-            default:
-                throw new PaserkException('Invalid version provided');
-        }
+        return match ($version) {
+            'k3' => new Version3(),
+            'k4' => new Version4(),
+            default => throw new PaserkException('Invalid version provided'),
+        };
     }
 
     /**
@@ -51,12 +42,6 @@ class Util
      */
     public static function getPaserkHeader(ProtocolInterface $pasetoVersion): string
     {
-        if ($pasetoVersion instanceof Version1) {
-            return 'k1';
-        }
-        if ($pasetoVersion instanceof Version2) {
-            return 'k2';
-        }
         if ($pasetoVersion instanceof Version3) {
             return 'k3';
         }
@@ -84,6 +69,7 @@ class Util
         } catch (SodiumException $ex) {
             $byref ^= $byref;
             unset($byref);
+            unset($ex); // We know what this is
         }
     }
 }
