@@ -8,8 +8,6 @@ use ParagonIE\Paserk\PaserkException;
 use ParagonIE\Paseto\Keys\AsymmetricSecretKey;
 use ParagonIE\Paseto\Keys\SymmetricKey;
 use ParagonIE\Paseto\Protocol\{
-    Version1,
-    Version2,
     Version3,
     Version4
 };
@@ -25,13 +23,11 @@ use PHPUnit\Framework\TestCase;
 class PBKWTest extends TestCase
 {
     /** @var ProtocolInterface[] */
-    protected $versions = [];
+    protected array $versions = [];
 
     public function setUp(): void
     {
         $this->versions = [
-            new Version1(),
-            new Version2(),
             new Version3(),
             new Version4(),
         ];
