@@ -7,8 +7,6 @@ use ParagonIE\Paserk\Operations\Key\SealingSecretKey;
 use ParagonIE\Paserk\Operations\PKE;
 use ParagonIE\Paseto\Keys\SymmetricKey;
 use ParagonIE\Paseto\Protocol\{
-    Version1,
-    Version2,
     Version3,
     Version4
 };
@@ -24,13 +22,11 @@ use PHPUnit\Framework\TestCase;
 class PKETest extends TestCase
 {
     /** @var ProtocolInterface[] */
-    protected $versions = [];
+    protected array $versions = [];
 
     public function setUp(): void
     {
         $this->versions = [
-            new Version1(),
-            new Version2(),
             new Version3(),
             new Version4()
         ];
