@@ -7,13 +7,12 @@ use ParagonIE\Paseto\Keys\{
     AsymmetricSecretKey
 };
 use ParagonIE\Paseto\Protocol\{
-    Version1,
-    Version2,
     Version3,
     Version4
 };
 use PHPUnit\Framework\TestCase;
 use ParagonIE\Paserk\Types\SecretType;
+use Exception;
 
 /**
  * Class PublicTest
@@ -23,32 +22,16 @@ use ParagonIE\Paserk\Types\SecretType;
  */
 class SecretTest extends TestCase
 {
-    /** @var AsymmetricPublicKey $v1pk */
-    protected $v1pk;
-    /** @var AsymmetricPublicKey $v2pk */
-    protected $v2pk;
-    /** @var AsymmetricPublicKey $v3pk */
-    protected $v3pk;
-    /** @var AsymmetricPublicKey $v4pk */
-    protected $v4pk;
-    /** @var AsymmetricSecretKey $v1sk */
-    protected $v1sk;
-    /** @var AsymmetricSecretKey $v2sk */
-    protected $v2sk;
-    /** @var AsymmetricSecretKey $v3sk */
-    protected $v3sk;
-    /** @var AsymmetricSecretKey $v4sk */
-    protected $v4sk;
+    protected AsymmetricPublicKey $v3pk;
+    protected AsymmetricPublicKey $v4pk;
+    protected AsymmetricSecretKey $v3sk;
+    protected AsymmetricSecretKey $v4sk;
 
     /**
-     * @throws \Exception
+     * @throws Exception
      */
     public function setUp(): void
     {
-        $this->v1sk = AsymmetricSecretKey::generate(new Version1());
-        $this->v1pk = $this->v1sk->getPublicKey();
-        $this->v2sk = AsymmetricSecretKey::generate(new Version2());
-        $this->v2pk = $this->v2sk->getPublicKey();
         $this->v3sk = AsymmetricSecretKey::generate(new Version3());
         $this->v3pk = $this->v3sk->getPublicKey();
         $this->v4sk = AsymmetricSecretKey::generate(new Version4());
@@ -58,24 +41,15 @@ class SecretTest extends TestCase
     public function testEncodeDecode()
     {
         /** @var AsymmetricPublicKey $key */
-        foreach ([$this->v1sk, $this->v2sk, $this->v3sk, $this->v4sk] as $key) {
+        foreach ([$this->v3sk, $this->v4sk] as $key) {
             $secret = new SecretType($key->getProtocol());
             $encoded = $secret->encode($key);
             $decoded = $secret->decode($encoded);
-            if ($key->getProtocol() instanceof Version1) {
-                // Compare raw -> compare PEM-encoded
-                $this->assertSame(
-                    $key->raw(),
-                    $decoded->raw(),
-                    'Key encoding failed: ' . $encoded
-                );
-            } else {
-                $this->assertSame(
-                    $key->encode(),
-                    $decoded->encode(),
-                    'Key encoding failed: ' . $encoded
-                );
-            }
+            $this->assertSame(
+                $key->encode(),
+                $decoded->encode(),
+                'Key encoding failed: ' . $encoded
+            );
         }
     }
 }
