@@ -4,7 +4,7 @@ namespace ParagonIE\Paserk\Operations\Key;
 
 use ParagonIE\Paserk\PaserkException;
 use ParagonIE\Paseto\Keys\AsymmetricPublicKey;
-use ParagonIE\Paseto\Protocol\Version1;
+use Exception;
 
 /**
  * Class SealingPublicKey
@@ -14,14 +14,12 @@ class SealingPublicKey extends AsymmetricPublicKey
 {
     /**
      * @return AsymmetricPublicKey
+     *
      * @throws PaserkException
+     * @throws Exception
      */
     public function toPasetoKey(): AsymmetricPublicKey
     {
-        if ($this->protocol instanceof Version1) {
-            throw new PaserkException("Version 1 keys cannot be converted!");
-        }
-
         return new AsymmetricPublicKey(
             $this->key,
             $this->protocol
