@@ -2,32 +2,25 @@
 declare(strict_types=1);
 namespace ParagonIE\Paserk\Types;
 
-use ParagonIE\ConstantTime\{
-    Base64,
-    Base64UrlSafe,
-    Binary
-};
 use ParagonIE\Paserk\{
     ConstraintTrait,
     PaserkException,
     PaserkTypeInterface,
     Util
 };
-use ParagonIE\Paseto\{
-    Exception\InvalidVersionException,
+use ParagonIE\Paseto\{Exception\InvalidVersionException,
+    Exception\PasetoException,
     KeyInterface,
     Keys\AsymmetricPublicKey,
-    Protocol\Version1,
     ProtocolCollection,
-    ProtocolInterface
-};
+    ProtocolInterface};
+use Exception;
+use SodiumException;
 use function
-    chunk_split,
     count,
     explode,
     hash_equals,
-    implode,
-    preg_replace;
+    implode;
 
 /**
  * Class PublicType
@@ -53,6 +46,7 @@ class PublicType implements PaserkTypeInterface
      * @return KeyInterface
      *
      * @throws PaserkException
+     * @throws Exception
      */
     public function decode(string $paserk): KeyInterface
     {
@@ -67,9 +61,6 @@ class PublicType implements PaserkTypeInterface
         $this->throwIfInvalidProtocol($version);
         /// @SPEC DETAIL: Algorithm Lucidity
 
-        if ($pieces[0] === 'k1') {
-            return $this->decodeV1($pieces[2]);
-        }
         return AsymmetricPublicKey::fromEncodedString(
             $pieces[2],
             $version
@@ -77,28 +68,11 @@ class PublicType implements PaserkTypeInterface
     }
 
     /**
-     * @param string $encoded
-     * @return AsymmetricPublicKey
-     * @throws \Exception
-     */
-    public function decodeV1(string $encoded): AsymmetricPublicKey
-    {
-        $raw = Base64UrlSafe::decode($encoded);
-        if (Binary::safeStrlen($raw) < 200) {
-            throw new PaserkException("Public key is too short");
-        }
-        $b64 = Base64::encode($raw);
-        $pem = '-----BEGIN PUBLIC KEY-----' . "\n" .
-            chunk_split($b64, 64, "\n") .
-            '-----END PUBLIC KEY-----';
-        return new AsymmetricPublicKey($pem, new Version1());
-    }
-
-    /**
      * @param KeyInterface $key
      * @return string
      *
      * @throws PaserkException
+     * @throws PasetoException
      */
     public function encode(KeyInterface $key): string
     {
@@ -109,39 +83,14 @@ class PublicType implements PaserkTypeInterface
         /// @SPEC DETAIL: Algorithm Lucidity
 
         $version = Util::getPaserkHeader($key->getProtocol());
-        switch ($version) {
-            case 'k1':
-                return implode('.', [
-                    $version,
-                    self::getTypeLabel(),
-                    $this->encodeV1($key->raw())
-                ]);
-            case 'k2':
-            case 'k3':
-            case 'k4':
-                return implode('.', [
-                    $version,
-                    self::getTypeLabel(),
-                    $key->encode()
-                ]);
-            default:
-                throw new PaserkException('Unknown version');
-        }
-    }
-
-    /**
-     * @param string $pk
-     * @return string
-     */
-    public function encodeV1(string $pk): string
-    {
-        $pem = preg_replace('#-{3,}(BEGIN|END) [^-]+-{3,}#', '', $pk);
-        $decoded = Base64::decode(preg_replace('#[^A-Za-z0-9+/]#', '', $pem));
-        $length = Binary::safeStrlen($decoded);
-        if ($length < 292) {
-            throw new PaserkException("Public key is too short: {$length}");
-        }
-        return Base64UrlSafe::encodeUnpadded($decoded);
+        return match ($version) {
+            'k3', 'k4' => implode('.', [
+                $version,
+                self::getTypeLabel(),
+                $key->encode()
+            ]),
+            default => throw new PaserkException('Unknown version'),
+        };
     }
 
     /**
@@ -157,8 +106,10 @@ class PublicType implements PaserkTypeInterface
      *
      * @param KeyInterface $key
      * @return string
+     *
      * @throws PaserkException
-     * @throws \SodiumException
+     * @throws PasetoException
+     * @throws SodiumException
      */
     public function id(KeyInterface $key): string
     {
