@@ -11,9 +11,8 @@ use ParagonIE\Paserk\Types\SecretType;
 use ParagonIE\Paseto\Exception\PasetoException;
 use ParagonIE\Paseto\Keys\AsymmetricSecretKey;
 use ParagonIE\Paseto\ProtocolInterface;
+use RangeException;
 use ParagonIE\Paseto\Protocol\{
-    Version1,
-    Version2,
     Version3,
     Version4
 };
@@ -23,16 +22,6 @@ use ParagonIE\Paseto\Protocol\{
  */
 class SecretTest extends KnownAnswers
 {
-    public function testV1()
-    {
-        $this->doJsonTest(new Version1(), 'k1.secret.json');
-    }
-
-    public function testV2()
-    {
-        $this->doJsonTest(new Version2(), 'k2.secret.json');
-    }
-
     public function testV3()
     {
         $this->doJsonTest(new Version3(), 'k3.secret.json');
@@ -43,11 +32,14 @@ class SecretTest extends KnownAnswers
         $this->doJsonTest(new Version4(), 'k4.secret.json');
     }
 
+    /**
+     * @param ProtocolInterface $version
+     * @param string $key
+     * @return AsymmetricSecretKey
+     * @throws Exception
+     */
     protected function getSecretKey(ProtocolInterface $version, string $key): AsymmetricSecretKey
     {
-        if ($version instanceof Version1) {
-            return new AsymmetricSecretKey($key, $version);
-        }
         return new AsymmetricSecretKey(Hex::decode($key), $version);
     }
 
@@ -65,7 +57,7 @@ class SecretTest extends KnownAnswers
                     try {
                         $secret->decode($test['paserk']);
                         $this->fail($test['name'] . ': ' . $test['comment']);
-                    } catch (ParserException | \RangeException | PasetoException | PaserkException $ex) {
+                    } catch (ParserException | RangeException | PasetoException | PaserkException $ex) {
                     }
                     continue;
                 }
@@ -74,22 +66,15 @@ class SecretTest extends KnownAnswers
                     $secretkey = $this->getSecretKey($version, $test['key']);
                     $secret->encode($secretkey);
                     $this->fail($test['name'] . ': '. $test['comment']);
-                } catch (ParserException | \RangeException | PasetoException | PaserkException $ex) {
+                } catch (ParserException | RangeException | PasetoException | PaserkException $ex) {
                 }
                 continue;
             }
             $secretkey = $this->getSecretKey($version, $test['key']);
-            if ($version instanceof Version1) {
-                $this->assertSame(
-                    $test['public-key'],
-                    $secretkey->getPublicKey()->raw()
-                );
-            } else {
-                $this->assertSame(
-                    $test['public-key'],
-                    Hex::encode($secretkey->getPublicKey()->raw())
-                );
-            }
+            $this->assertSame(
+                $test['public-key'],
+                Hex::encode($secretkey->getPublicKey()->raw())
+            );
             $this->assertSame($test['paserk'], $secret->encode($secretkey), $test['name']);
         }
     }
