@@ -5,12 +5,12 @@ namespace ParagonIE\Paserk\Tests\KAT;
 use ParagonIE\ConstantTime\Hex;
 use ParagonIE\Paserk\Operations\Wrap;
 use ParagonIE\Paserk\Operations\Wrap\Pie;
-use ParagonIE\Paserk\Types\SecretWrap;
+use ParagonIE\Paserk\PaserkException;
 use ParagonIE\Paserk\Tests\KnownAnswers;
+use ParagonIE\Paserk\Types\SecretWrap;
 use ParagonIE\Paseto\Keys\SymmetricKey;
+use Throwable;
 use ParagonIE\Paseto\Protocol\{
-    Version1,
-    Version2,
     Version3,
     Version4
 };
@@ -21,16 +21,6 @@ use ParagonIE\Paseto\ProtocolInterface;
  */
 class SecretWrapPieTest extends KnownAnswers
 {
-    public function testV1()
-    {
-        $this->doJsonTest(new Version1(), 'k1.secret-wrap.pie.json');
-    }
-
-    public function testV2()
-    {
-        $this->doJsonTest(new Version2(), 'k2.secret-wrap.pie.json');
-    }
-
     public function testV3()
     {
         $this->doJsonTest(new Version3(), 'k3.secret-wrap.pie.json');
@@ -45,6 +35,8 @@ class SecretWrapPieTest extends KnownAnswers
      * @param ProtocolInterface $version
      * @param string $name
      * @param array $tests
+     *
+     * @throws PaserkException
      */
     protected function genericTest(ProtocolInterface $version, string $name, array $tests): void
     {
@@ -54,22 +46,17 @@ class SecretWrapPieTest extends KnownAnswers
             if ($test['expect-fail']) {
                 try {
                     $wrapper->decode($test['paserk']);
-                } catch (\Throwable $exception) {
+                } catch (Throwable $exception) {
                     continue;
                 }
                 $this->fail($name . ' > ' . $test['name'] . ': '. $test['comment']);
             }
             $unwrapped = $wrapper->decode($test['paserk']);
-
-            if ($version instanceof Version1) {
-                $this->assertSame($test['unwrapped'], $unwrapped->raw(), $test['name']);
-            } else {
-                $this->assertSame(
-                    $test['unwrapped'],
-                    Hex::encode($unwrapped->raw()),
-                    $test['name']
-                );
-            }
+            $this->assertSame(
+                $test['unwrapped'],
+                Hex::encode($unwrapped->raw()),
+                $test['name']
+            );
         }
     }
 }
