@@ -2,17 +2,19 @@
 declare(strict_types=1);
 namespace ParagonIE\Paserk\Tests\KAT;
 
+use Exception;
 use FG\ASN1\Exception\ParserException;
 use ParagonIE\ConstantTime\Hex;
 use ParagonIE\Paserk\PaserkException;
 use ParagonIE\Paserk\Tests\KnownAnswers;
 use ParagonIE\Paserk\Types\Sid;
+use ParagonIE\Paseto\Exception\InvalidVersionException;
 use ParagonIE\Paseto\Exception\PasetoException;
 use ParagonIE\Paseto\Keys\AsymmetricSecretKey;
 use ParagonIE\Paseto\ProtocolInterface;
+use RangeException;
+use SodiumException;
 use ParagonIE\Paseto\Protocol\{
-    Version1,
-    Version2,
     Version3,
     Version4
 };
@@ -22,16 +24,6 @@ use ParagonIE\Paseto\Protocol\{
  */
 class SidTest extends KnownAnswers
 {
-    public function testV1()
-    {
-        $this->doJsonTest(new Version1(), 'k1.sid.json');
-    }
-
-    public function testV2()
-    {
-        $this->doJsonTest(new Version2(), 'k2.sid.json');
-    }
-
     public function testV3()
     {
         $this->doJsonTest(new Version3(), 'k3.sid.json');
@@ -42,14 +34,27 @@ class SidTest extends KnownAnswers
         $this->doJsonTest(new Version4(), 'k4.sid.json');
     }
 
+    /**
+     * @param ProtocolInterface $version
+     * @param string $key
+     * @return AsymmetricSecretKey
+     *
+     * @throws Exception
+     */
     protected function getSecretKey(ProtocolInterface $version, string $key): AsymmetricSecretKey
     {
-        if ($version instanceof Version1) {
-            return new AsymmetricSecretKey($key, $version);
-        }
         return new AsymmetricSecretKey(Hex::decode($key), $version);
     }
 
+    /**
+     * @param ProtocolInterface $version
+     * @param string $name
+     * @param array $tests
+     *
+     * @throws PaserkException
+     * @throws InvalidVersionException
+     * @throws SodiumException
+     */
     protected function genericTest(ProtocolInterface $version, string $name, array $tests): void
     {
         foreach ($tests as $test) {
@@ -58,7 +63,7 @@ class SidTest extends KnownAnswers
                     $secretkey = $this->getSecretKey($version, $test['key']);
                     Sid::encodeSecret($secretkey);
                     $this->fail($test['name'] . ': '. $test['comment']);
-                } catch (ParserException | \RangeException | PasetoException | PaserkException $ex) {
+                } catch (ParserException | RangeException | PasetoException | PaserkException $ex) {
                 }
                 continue;
             }
