@@ -2,16 +2,20 @@
 declare(strict_types=1);
 namespace ParagonIE\Paserk\Tests\Types;
 
-use ParagonIE\Paserk\Types\Lid;
-use ParagonIE\Paserk\Types\LocalWrap;
+use Exception;
+use ParagonIE\Paserk\{
+    PaserkException,
+    Types\Lid,
+    Types\LocalWrap
+};
+use ParagonIE\Paseto\Exception\InvalidVersionException;
 use ParagonIE\Paseto\Keys\SymmetricKey;
 use ParagonIE\Paseto\Protocol\{
-    Version1,
-    Version2,
     Version3,
     Version4
 };
 use PHPUnit\Framework\TestCase;
+use SodiumException;
 
 /**
  * Class LocalWrapTest
@@ -21,24 +25,27 @@ use PHPUnit\Framework\TestCase;
  */
 class LocalWrapTest extends TestCase
 {
+    protected SymmetricKey $v3key;
+    protected SymmetricKey $v4key;
 
-    protected $v1key;
-    protected $v2key;
-    protected $v3key;
-    protected $v4key;
-
+    /**
+     * @throws Exception
+     */
     public function setUp(): void
     {
-        $this->v1key = SymmetricKey::generate(new Version1());
-        $this->v2key = SymmetricKey::generate(new Version2());
         $this->v3key = SymmetricKey::generate(new Version3());
         $this->v4key = SymmetricKey::generate(new Version4());
     }
 
+    /**
+     * @throws PaserkException
+     * @throws InvalidVersionException
+     * @throws SodiumException
+     */
     public function testWrap()
     {
         /** @var SymmetricKey $key */
-        foreach ([$this->v1key, $this->v2key, $this->v3key, $this->v4key] as $key) {
+        foreach ([$this->v3key, $this->v4key] as $key) {
             // Generate wrapping key
             $version = $key->getProtocol();
             $wk = SymmetricKey::generate($version);
